chore: 폴더 변경 및 유틸 함수 폴더 분리
