+++ src/app/timetable/components/Slot.module.scss
