+++ src/app/timetable/components/Timetable.module.scss
