+++ src/app/timetable/components/Timetable.type.ts
