+++ src/app/timetable/constants/index.ts
