+++ src/app/timetable/mocks/timetableMockData.ts
@@ -56,4 +56,47 @@ const taskList = [
   },
 ];
 
+const duplicatedTimeTaskList = [
+  {
+    id: 1,
+    title: 'title1',
+    subTitle: 'subTitle1',
+    slotColor: 'red',
+    startTime: getDateFromTime(12, 0, 0),
+    endTime: getDateFromTime(13, 0, 0),
+  },
+  {
+    id: 2,
+    title: 'title2',
+    subTitle: 'subTitle2',
+    slotColor: 'blue',
+    startTime: getDateFromTime(12, 10, 0),
+    endTime: getDateFromTime(12, 30, 0),
+  },
+  {
+    id: 3,
+    title: 'title3',
+    subTitle: 'subTitle3',
+    slotColor: 'purple',
+    startTime: getDateFromTime(18, 0, 0),
+    endTime: getDateFromTime(20, 0, 0),
+  },
+  {
+    id: 5,
+    title: '02:30',
+    subTitle: '02:50',
+    slotColor: 'green',
+    startTime: getDateFromTime(17, 30, 0),
+    endTime: getDateFromTime(18, 30, 0),
+  },
+  {
+    id: 4,
+    title: '01:00',
+    subTitle: '02:20',
+    slotColor: 'brown',
+    startTime: getDateFromTime(19, 30, 0),
+    endTime: getDateFromTime(21, 20, 0),
+  },
+];
+
 export { startTime, endTime, taskList, slotTime, height };
