+++ src/app/timetable/utils/timeTable.util.test.ts
